Removed `dynatos::dyn_element`.
It was a bad approach to a problem, and had edge cases relating to what happened to the returned node if modified.
In the end, the only correct way to do it was to append it as a child and not touch it anymore, although `with_dyn_child` exists for that.